Update projects to .NET 8
When we brought in Microsoft.Extensions.Caching.StackExchangeRedis, it bumped the transitive version of Microsoft.IdentityModel to v7+ while before we used v6.x. There was a breaking change between these versions that switched Json.NET to System.Text.Json. However, our dependencies still used v6 of the OpenIdConnect which at runtime cannot find the new type.

Easiest way to fix this is to bring in .NET 8 so the versions are all on the latest supported platforms. This brings in some required changes for downstream API access and application insights configuration.

See https://github.com/AzureAD/azure-activedirectory-identitymodel-extensions-for-dotnet/wiki/IdentityModel-7x#serialization-issues for details.