Add error handling during the creation of Entra ID App Registration (#398)
If this command to save a client_secret were to fail then the script
would exit with failure. The result would prevent the reader from being
able to create the 2nd App Registration which is required to
successfully boot strap the web api project.

With this update:
- Flow remains unchanged if there is no error.
- If there is an error it will be treated as a warning. The impact to
the reader is that they would only be able to perform unauthenticated
actions until the client_secret is saved to Key Vault.

---------

Co-authored-by: Taylor Southwick <tasou@microsoft.com>